:recycle: Refactor `byte_offset_to_line_column` to a new `positions.rs` file under `parser/src/common/`
