+++ parser/src/test_utils.rs
@@ -1,24 +1,8 @@
 // Test support utilities (only compiled during tests)
 // Shared helpers to reduce duplication across lexer/parser tests.
 
-/// Convert a byte offset into (line, column) 1-based coordinates.
-/// Counts Unicode scalar values; advances line on '\n' only.
-pub fn byte_offset_to_line_column(text: &str, offset: usize) -> (usize, usize) {
-    let mut line = 1;
-    let mut column = 1;
-    for (i, ch) in text.char_indices() {
-        if i >= offset {
-            break;
-        }
-        if ch == '\n' {
-            line += 1;
-            column = 1;
-        } else {
-            column += 1;
-        }
-    }
-    (line, column)
-}
+// Re-export public utility for convenience in tests.
+pub use crate::common::position::byte_offset_to_line_column;
 
 /// Collect stringified context frames from a winnow ParseError.
 pub fn collect_context<I>(
